refactor root context builder into root context
